refactor: ListPage 중복되는 코드 분리
